third_party/blink/web_tests/platform/generic/fast/loader/scroll-restore-overrides-fragment-expected.txt

Tests that after a navigation, scroll restoration from history takes precedence over scrolling to the hash fragment.

PASS 600 is 600
PASS successfullyParsed is true

TEST COMPLETE
